Merging Data Visualization into master branch (#12)
* Add files via upload

* Delete DAS_Project.Rmd

* Delete DAS_GLM.qmd

* Delete dataset15.csv

* Update data visulisation.qmd (#3)

* retain the data exhibit 0 in aroma (#5)

compared with the code before, i retain the data with 0 in aroma, actually I  am  struggle with it. Cuz only one data exhibit 0 in aroma flavor and acidity. How to deal with this data is more appropriate?

* Update data visulisation.qmd (#10)

Co-authored-by: bibabobibabo <162199841+bibabobibabo@users.noreply.github.com>

* Update data visulisation.qmd (#11)

Added comments on code and added correlation matrix and heatmap

Co-authored-by: hancyyyy <161476952+hancyyyy@users.noreply.github.com>

* Rename data visulisation.qmd to DAS_Project.qmd

changed file name to merge into the master branch

---------

Co-authored-by: bibabobibabo <162199841+bibabobibabo@users.noreply.github.com>
Co-authored-by: hancyyyy <161476952+hancyyyy@users.noreply.github.com>